Problem stumbled to: Paralelism wasn't working as a single session can only have a certain ammount (5?) queryies simutaneously
Solution:

Use the provided scoped_session #https://docs.sqlalchemy.org/en/13/orm/contextual.html
Was also added for the possibility to work with context (with statement), in order to simplify handling it.

Possible future problems:
There is a limit for the ammount of simutaneous sessions. Using pool_size argument on create_engine function might be the solution #https://stackoverflow.com/questions/24956894/sql-alchemy-queuepool-limit-overflow
(I actually stumbled on that problem already, which is why I implemented the use of context)